Codechange: Split types out of newgrf_text.h
