+++ src/newgrf_text.h
@@ -10,24 +10,9 @@
 #ifndef NEWGRF_TEXT_H
 #define NEWGRF_TEXT_H
 
-#include "string_type.h"
+#include "newgrf_text_type.h"
 #include "strings_type.h"
 #include "table/control_codes.h"
-#include <utility>
-
-/** This character, the thorn ('þ'), indicates a unicode string to NFO. */
-static const char32_t NFO_UTF8_IDENTIFIER = 0x00DE;
-
-/** A GRF text with associated language ID. */
-struct GRFText {
-	uint8_t langid;      ///< The language associated with this GRFText.
-	std::string text; ///< The actual (translated) text.
-};
-
-/** A GRF text with a list of translations. */
-typedef std::vector<GRFText> GRFTextList;
-/** Reference counted wrapper around a GRFText pointer. */
-typedef std::shared_ptr<GRFTextList> GRFTextWrapper;
 
 StringID AddGRFString(uint32_t grfid, uint16_t stringid, uint8_t langid, bool new_scheme, bool allow_newlines, std::string_view text_to_add, StringID def_string);
 StringID GetGRFStringID(uint32_t grfid, StringID stringid);
@@ -48,28 +33,5 @@ void StopTextRefStackUsage();
 bool UsingNewGRFTextStack();
 struct TextRefStack *CreateTextRefStackBackup();
 void RestoreTextRefStackBackup(struct TextRefStack *backup);
-
-/** Mapping of language data between a NewGRF and OpenTTD. */
-struct LanguageMap {
-	/** Mapping between NewGRF and OpenTTD IDs. */
-	struct Mapping {
-		uint8_t newgrf_id;  ///< NewGRF's internal ID for a case/gender.
-		uint8_t openttd_id; ///< OpenTTD's internal ID for a case/gender.
-	};
-
-	/* We need a vector and can't use SmallMap due to the fact that for "setting" a
-	 * gender of a string or requesting a case for a substring we want to map from
-	 * the NewGRF's internal ID to OpenTTD's ID whereas for the choice lists we map
-	 * the genders/cases/plural OpenTTD IDs to the NewGRF's internal IDs. In this
-	 * case a NewGRF developer/translator might want a different translation for
-	 * both cases. Thus we are basically implementing a multi-map. */
-	std::vector<Mapping> gender_map; ///< Mapping of NewGRF and OpenTTD IDs for genders.
-	std::vector<Mapping> case_map;   ///< Mapping of NewGRF and OpenTTD IDs for cases.
-	int plural_form;                 ///< The plural form used for this language.
-
-	int GetMapping(int newgrf_id, bool gender) const;
-	int GetReverseMapping(int openttd_id, bool gender) const;
-	static const LanguageMap *GetLanguageMap(uint32_t grfid, uint8_t language_id);
-};
 
 #endif /* NEWGRF_TEXT_H */
